Implemented Log +  new test is added in main.cpp
Tip: If the log.txt file isn't found in the main directory, it shall be stored inside the cmake-build-debug folder.
The tip applies to memory.txt folder as well.

+++ m_stdio.h
@@ -35,7 +35,7 @@ unsigned char* m_strncat(unsigned char* dest,size_t index,unsigned char* src,siz
     }
     return dest;
 }
-int m_strncmp(unsigned char* str1,unsigned char* str2,size_t num){
+int m_strncmp(unsigned char* str1,unsigned char* str2,int num){
     if(!str1 || !str2){
         return 5;
     }
